Improve meta track syncing
