+++ src/lib/components/track-meta.svelte
@@ -2,7 +2,6 @@
 	import {pg} from '$lib/db'
 	import {searchMusicBrainz} from '$lib/sync/musicbrainz'
 	import {logger} from '$lib/logger'
-	import {onMount} from 'svelte'
 	const log = logger.ns('<TrackMeta>').seal()
 
 	/**
@@ -13,28 +12,30 @@
 	const {track} = $props()
 
 	let loading = $state(false)
-	let error = $state(null)
-	let results = $state(null)
+	let error = $state('')
+	let musicbrainz_data = $state(null)
 
 	const ytid = $derived(track?.youtube_data.id)
 
-	$effect(() => {
+	$effect(async () => {
 		if (!ytid) return
 		loading = true
-		error = null
-		results = null
 
-		searchMusicBrainz(track.title).then(async (result) => {
-			if (result) {
-				results = result
-				await pg.sql`update track_meta set updated_at = CURRENT_TIMESTAMP, musicbrainz_updated_at = CURRENT_TIMESTAMP, musicbrainz_data = ${JSON.stringify(result)} where ytid = ${ytid}`
-				log.log('musicbrainz_data:updated', track.title, result)
-			} else {
-				error = 'No matches found'
-				results = null
-			}
-			loading = false
-		})
+		musicbrainz_data = await searchMusicBrainz(track.title)
+		if (musicbrainz_data) {
+			error = null
+			await pg.sql`
+					update track_meta 
+					set updated_at = CURRENT_TIMESTAMP,
+					musicbrainz_updated_at = CURRENT_TIMESTAMP,
+					musicbrainz_data = ${JSON.stringify(musicbrainz_data)}
+					where ytid = ${ytid}`
+			log.log('musicbrainz_data:updated', track.title, musicbrainz_data)
+		} else {
+			error = 'No matches found'
+			musicbrainz_data = null
+		}
+		loading = false
 	})
 
 	$effect(async () => {
@@ -53,76 +54,12 @@
 </script>
 
 {#if loading}
-	<p>Loading MusicBrainz metadata...</p>
+	<p>Loading meta data from YouTube and MusicBrainz...</p>
 {:else if error}
 	<p>Error: {error}</p>
 {:else if results}
-	{@const recording = results.recording}
-	{@const artist = recording['artist-credit']?.map((ac) => ac.name).join(', ')}
-	{@const duration = recording.length
-		? `${Math.floor(recording.length / 60000)}:${String(Math.floor((recording.length % 60000) / 1000)).padStart(2, '0')}`
-		: null}
-	{@const releaseDate = recording['first-release-date']}
-
-	{@const primaryRelease =
-		recording.releases?.find(
-			(r) =>
-				r.status === 'Official' &&
-				r['release-group']?.['primary-type'] === 'Album' &&
-				!r['release-group']?.['secondary-types']?.includes('Compilation')
-		) ||
-		recording.releases?.find(
-			(r) => r.status === 'Official' && r['release-group']?.['primary-type'] === 'EP'
-		) ||
-		recording.releases?.[0]}
-
-	<div>
-		<p>
-			Please note, we searched MusicBrainz for the track title. This might not be the track you are
-			looking for. But it also might be.
-		</p>
-		<p>
-			MusicBrainz ID: {recording.id}<br />
-			recording title: {recording.title}<br />
-			artist: {artist}<br />
-			duration: {duration}<br />
-			primary release: {primaryRelease.title}<br />
-			release date: {releaseDate}<br />
-			release date: {new Date(releaseDate).toLocaleDateString()}<br />
-			tags: {recording.tags?.map((tag) => tag.name).join(', ')}<br />
-		</p>
-
-		{#if recording.releases && recording.releases.length > 1}
-			<details>
-				<summary><strong>All Releases ({recording.releases.length})</strong></summary>
-				<ul>
-					{#each recording.releases as release (release.id)}
-						<li>
-							<strong>{release.title}</strong>
-							{#if release.date}({release.date}){/if}
-							- {release.status}
-							{#if release['release-group']?.['primary-type']}
-								{release['release-group']['primary-type']}
-							{/if}
-							{#if release['release-group']?.['secondary-types']}
-								({release['release-group']['secondary-types'].join(', ')})
-							{/if}
-						</li>
-					{/each}
-				</ul>
-			</details>
-		{/if}
-
-		<details>
-			<summary><strong>Search Debug</strong></summary>
-			<p>
-				original: {results.originalTitle}<br />
-				cleaned: {results.parsed.cleaned}<br />
-				strategy: {results.searchDescription}<br />
-				score: {recording.score}%
-			</p>
-		</details>
-	</div>
+	<p>found soemthing</p>
+	<pre>{JSON.stringify(results, null, 2)}</pre>
 {:else}
 	<p>No MusicBrainz metadata found for "{track.title}"</p>
 {/if}
